% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_shubham_c}
\alias{scale_color_shubham_c}
\title{Continuous Color Scale for ggplot2}
\usage{
scale_color_shubham_c(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating the palette type.
Options are "main" or "grays"}

\item{reverse}{Logical. If TRUE, reverses the order of colors}

\item{...}{Additional arguments passed to ggplot2::scale_color_gradientn()}
}
\value{
A ggplot2 color scale
}
\description{
Creates a continuous color scale for ggplot2 using Shubham's color palettes
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_shubham_c()
}
