% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_shubham_c}
\alias{scale_fill_shubham_c}
\title{Continuous Fill Scale for ggplot2}
\usage{
scale_fill_shubham_c(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating the palette type.
Options are "main" or "grays"}

\item{reverse}{Logical. If TRUE, reverses the order of colors}

\item{...}{Additional arguments passed to ggplot2::scale_fill_gradientn()}
}
\value{
A ggplot2 fill scale
}
\description{
Creates a continuous fill scale for ggplot2 using Shubham's color palettes
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_shubham_c()
}
