% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_shubham_d}
\alias{scale_fill_shubham_d}
\title{Discrete Fill Scale for ggplot2}
\usage{
scale_fill_shubham_d(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating the palette type.
Options are "main" or "grays"}

\item{reverse}{Logical. If TRUE, reverses the order of colors}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\value{
A ggplot2 fill scale
}
\description{
Creates a discrete fill scale for ggplot2 using Shubham's color palettes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(am))) +
  geom_bar() +
  scale_fill_shubham_d()
}
