% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{shubham_cols}
\alias{shubham_cols}
\title{Shubham Colors}
\format{
A list of color hex codes
}
\usage{
shubham_cols
}
\description{
A collection of shubhamd colors for use in plots and visualizations
}
\keyword{datasets}
