% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{shubham_pal_c}
\alias{shubham_pal_c}
\title{Create a Continuous Color Palette Function}
\usage{
shubham_pal_c(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating the palette type.
Options are "main" or "grays"}

\item{reverse}{Logical. If TRUE, reverses the order of colors}

\item{...}{Additional arguments passed to grDevices::colorRampPalette()}
}
\value{
A function that takes an integer argument (n) and returns n colors
}
\description{
Creates a function that returns a continuous color palette from
Shubham's color schemes
}
\examples{
# Get 100 colors interpolated from main palette
pal <- shubham_pal_c()
pal(100)
}
