% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shubhamPalette-package.R
\docType{package}
\name{shubhamPalette-package}
\alias{shubhamPalette}
\alias{shubhamPalette-package}
\title{shubhamPalette: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
