% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shubhamPalette-package.R
\docType{package}
\name{shubhamPalette-package}
\alias{shubhamPalette}
\alias{shubhamPalette-package}
\title{shubhamPalette: Custom Palettes and Themes}
\description{
Custom palettes and ggplot2 themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/shubhamPalette}
  \item Report bugs at \url{https://github.com/shubhamdutta26/shubhamPalette/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
